feat(float80 byte to uint64 conversion)

+++ pkg/codec/audio/audio.go
@@ -2,7 +2,7 @@ package audio
 
 type Audio struct {
 	NumChannels uint16
-	SampleRate  uint32
+	SampleRate  uint64
 	BitDepth    uint16
 	BitRate     uint32
 	Samples     []byte
